refactor(PAN-3049): extract duplicate-stack doctor check to its own module
The previous commit added the duplicate-compose-stack check inline in
doctor.ts, pushing it from 998 to 1102 lines and past the 1000-line
god-file ceiling (scripts/lint-file-size.sh), which blocked the push.

Extracts diagnoseDuplicateComposeStacks and checkDuplicateComposeStacks
into src/cli/commands/doctor-duplicate-stacks.ts, mirroring the existing
doctor-inotify.ts / doctor-state-worktree.ts pattern (a locally-redeclared
CheckResult interface avoids a module cycle with doctor.ts). doctor.ts now
only imports and wires the check in, landing at exactly 1000 lines.
